Refactored lens profile publish issue
